Bug Fixes (#1271)
* Fix typo in SetValueNumber parameter name

Renamed the parameter valueVumber to valueNumber in the SetValueNumber method and updated all references accordingly. This improves code clarity and prevents confusion from the previous typo.

* Add null checks in IsPhiUseless for value numbers

Prevent null reference exceptions in IsPhiUseless by checking for null value numbers before comparison. Ensure all operands have valid value numbers before determining if a Phi node is useless.

* Fix Info64 calculation, null checks, and page property

- Correct Info64 property in relocation entries to use ulong and proper bitwise operations for 64-bit support.
- Add null check for context.Operand1 in IsPlugged to prevent exceptions.
- Fix IsPreviousPage to check Previous instead of Next for accuracy.

* Fix AlignUp(int) to call AlignUp, not AlignDown

Previously, AlignUp(int, int) incorrectly called AlignDown, causing it to align values downward. This commit corrects the method to call AlignUp((uint), (uint)), ensuring proper upward alignment for integer inputs.

* Add general guideline to avoid unsafe code; fix grammar

Added a "General Guidelines" section to copilot-instructions.md instructing not to propose or include unsafe code blocks or APIs. Also corrected grammar in the "@azure Rule - Enable Best Practices" guideline by adding a comma.

+++ Source/Mosa.Compiler.Framework/Linker/Elf/RelocationAddendEntry.cs
@@ -34,7 +34,7 @@ public class RelocationAddendEntry
 	/// <summary>
 	/// Gets the Info field for 64bit elf
 	/// </summary>
-	public ushort Info64 => (ushort)((Symbol << 32) | ((ushort)RelocationType & 0xFFFFFFFF));
+	public ulong Info64 => (ulong)((Symbol << 32) | ((ulong)RelocationType & 0xFFFFFFFF));
 
 	public static uint GetEntrySize(LinkerFormatType elfType)
 	{
